Move path manipulation function to utils
