+++ libs/remote_storage/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2021"
 anyhow = { version = "1.0", features = ["backtrace"] }
 async-trait = "0.1"
 metrics = { version = "0.1", path = "../metrics" }
+utils = { version = "0.1", path = "../utils" }
 once_cell = "1.13.0"
 rusoto_core = "0.48"
 rusoto_s3 = "0.48"
